Corrigindo de float para decimal

+++ database/migrations/2022_01_10_230738_create_taxas_table.php
@@ -15,7 +15,7 @@ class CreateTaxasTable extends Migration
     {
         Schema::create('taxas', function (Blueprint $table) {
             $table->id();
-            $table->float('taxa', 10, 2)->default(0);
+            $table->decimal('taxa', 10, 2)->default(0);
             $table->unsignedBigInteger('user_id');
             $table->timestamps();
